Adding logic for mouse following focus.
When focus changes, SIWindow queries the Configuration to see if the user would
like for their mouse cursor to follow the focus change.  If so, it will position
the cursor in the midpoint of the window.  The default setting is for this not
to happen.

Also, tidied up the logic for ignoreMenuBar to be consistent with the rest of
the codebase.