Cleaning up mouse follows focus logic.
Instead of extending the SIWindow class with a method of the same name,
I added a new method `AMFocusWindow` which AMWindowManager now calls.
This method calls the original `focusWindow` method of SIWindow
and then performs mouse follows focus if it is configured.  Also added
a brief README entry for the setting.