Ensure active windows are on current space
Under some circumstances (the major one seems to be Mission Control), windows
from many spaces can be on screen according to CoreGraphics. With the addition
of a private API we can actually determine the space a window should be on and
use that to verify against what the current space is.

This may actually fix #254