Add shortcuts to move focused window left/right a space
This changeset adds two shortcuts:
 * `mod2 + left`, and
 * `mod2 + right`

These act in a similar way to the existing `mod2 + [n]`, but in a relative, not absolute, fashion.

The behaviour mirrors symmetrically the OSX default `ctrl + [n]` vs. `ctrl + left`/`ctrl + right`
for space switching. The new keybindings are unlikely to clash, fit logically with existing
Amethyst/System keybindings, are idiomatic, and natural to use.

The Spaces API seems to be undocuemnted/reverse-engineered, so I had to copy existing Amethyst
code.

I added four convenience functions for getting the current space, and for getting the number
of configured spaces.  These can be supplied a screen, or try to find the "currently focused" one.

Otherwise, I followed the implementation of `pushFocusedWindowToSpace`, and updated the README.

N.B. I return -1 from certain functions to signal failure.  Is this idiomatic Objective C?
N.B. I freely cast between `NSUIntger` and `CGSSpace`.  Is this acceptable?

I am unable to test the code, so __please compile and run the code before merging it anywhere
important!__