Refactor a bunch of classes to reduce complexity
The worst offender here was WindowManager, which has been split up into:

* `WindowModifier` - responsible for the alteration of window state.
* `FocusFollowsMouseManager` - responsible for managing focus follows mouse behavior.
* Some behavior has moved into extensions for `SIWindow` and `NSScreen`.

Additionally, `UserConfiguration` has been broken up as well. It is not largely responsible for the reading and writing of preferences. Hot key handler construct has been moved to the `HotKeyHandler` and a `LayoutManager` has been factored out to handle all of the layout class stuff.