Keep windows centered when adjusting for minimum size
Minimum height and width settings were adding height or width to the
bottom or left of the window. Focused windows get adjusted to fit inside
the screen, but they were not getting adjusted up/left such as to keep
the windows centered if they were not touching a corner. This change
makes sure they are centered, and also keeps them inside the bounds of
the screen on the left side of the screen.