Simplify layout structures
The interactions were kind of a mess. The screen manager had to pass the window activity cache through, like, four layers of indirection and a handful of protocols. I decided to simplify this to just be the screen manager creates a reflow operation and passes in a layout, the layout generates frame assignments, and then the reflow operation applies the frames. This maintains the testability of isolating layout logic from actual application logic, but doesn't require so much passing around.