Fix LayoutsPreferences Auto Layout Constraints
Several of the layout constraints were anchored to the table view for
the list of selected layouts instead of the scroll view containing the
table. Once the list had enough elements to cause scrolling the
constraints became unsatisfiable because the "+" button was constrainted
to be below the table view but also a fixed distance from the top of the
window. This causes the UI to be completely unusable as the list view
disappears completely.