Remove redundant constraint in LayoutsPreferences
The layout constraint from the top of the view to the top of the "+"
square button was redundant and contributed to the layout becoming
unsatisfiable. The constraints from the top of the view to "General" to
"Window Resize Interval" to "Horizontal Line" to "Bordered Scroll View"
to the + "Square" already give a fixed vertial position to the element.