Throw window to any space across multiple screens.
Currently throwing windows to spaces given the space number only works
for a single screen. You cannot throw a window to a space belonging to a
different screen. This is confusing when managing windows in a
multi-screen setup as you need to think about screen numbers locally to
the screen as well as globally across all screens. It creates more steps
when issuing keyboard commands i.e. instead of throwing a window to
space X on which exists on screen A directly, you must first throw the
window to screen A and then throw it to space X' which is the index of
screen X local to screen A. Finally the current behaviour deviates from
the existing XMonad behaviour which allows throwing windows to arbitrary
globally indexed workspaces.